Use an unsigned integer for representing the PIN length
The value 0 is not allowed by the HCI spec so it can be used to denote
"not applicable" or "unknown". Also, the management interface will use
an unsigned integer for the PIN length so mirroring that in the core
daemon code makes sense. This patch also takes care of handling
previously stored -1 values and converts them to 0 whenever they are
detected.

+++ src/adapter.h
@@ -69,7 +69,7 @@ struct link_key_info {
 	bdaddr_t bdaddr;
 	unsigned char key[16];
 	uint8_t type;
-	int pin_len;
+	uint8_t pin_len;
 };
 
 struct remote_dev_info {
